WIP: MP4 Builder improvements
