+++ lib/mp4/atoms/atom-mp4a.js
@@ -51,11 +51,14 @@ class AtomMP4A extends Atom {
             let size = buffer.readUInt32BE(offset);
             offset += 4;
             let type = buffer.toString('ascii', offset, offset + 4);
+            console.log(size, type);
             offset += 4;
             if (size === 0) {
                 break;
             }
             if (type === Utils.ATOM_ESDS) {
+                // Parse Elementary Stream Descriptor
+                console.log(buffer[offset], buffer[offset + 1], buffer[offset + 2], buffer[offset + 3], buffer[offset + 4]);
                 offset += 5;
                 offset += readSize(buffer, offset).read;
                 offset += 4;
@@ -86,15 +89,15 @@ class AtomMP4A extends Atom {
         // sample size
         buffer.writeUInt16BE(this.sampleSize, offset + 26);
         // sample rate
-        buffer.writeUInt32BE(this.sampleRate, offset + 32);
+        buffer.writeUInt32BE(this.sampleRate, offset + 30);
         // extra data
-        buffer.writeUInt32BE(this.extraData.length + 4, offset + 38);
-        buffer.write(Utils.ATOM_ESDS, offset + 42);
-        this.extraData.copy(buffer, offset + 46, 4);
+        buffer.writeUInt32BE(this.extraData.length + 4, offset + 36);
+        buffer.write(Utils.ATOM_ESDS, offset + 40);
+        this.extraData.copy(buffer, offset + 44, 4);
     }
 
     bufferSize() {
-        return 42 + this.extraData.length;
+        return 40 + this.extraData.length;
     }
 
 }
