+++ lib/mp4/parser-impl.js
@@ -137,6 +137,7 @@ class ParserImpl {
 
         track.duration = mdhdAtom.duration;
         track.timescale = mdhdAtom.timescale;
+        console.log(track.extraData);
         let codecInfo = CodecParser.parse(track.extraData);
         if (codecInfo !== null) {
             track.codec = codecInfo.codec();
