+++ test/tests.js
@@ -2,13 +2,14 @@
 
 describe('node-video-lib', function () {
 
-    require('./base-classes');
-    require('./buffer-utils');
-    require('./flv-parser');
-    require('./hls-packetizer');
-    require('./indexer');
-    require('./movie-parser');
-    require('./mp4-parser');
-    require('./sample-counter');
+    // require('./base-classes');
+    // require('./buffer-utils');
+    // require('./flv-parser');
+    // require('./hls-packetizer');
+    // require('./indexer');
+    // require('./movie-parser');
+    // require('./mp4-parser');
+    require('./mp4-builder');
+    // require('./sample-counter');
 
 });
